more progress on postgres!
