--- mongo-docker-compose.yml
@@ -1,25 +1,0 @@
-services:
-  mongodb:
-    image: mongodb/mongodb-community-server:6.0.6-ubuntu2204
-    container_name: mongodb
-    restart: unless-stopped
-    environment:
-      MONGODB_INITDB_ROOT_USERNAME: user
-      MONGODB_INITDB_ROOT_PASSWORD: password
-    ports:
-      - 27017:27017
-
-# Disable mongo express, using compass instead
-#  mongo-express:
-#    image: mongo-express:1.0.0-20-alpine3.18
-#    container_name: mongo-express
-#    restart: unless-stopped
-#    ports:
-#      - 8081:8081
-#    environment:
-#      ME_CONFIG_BASICAUTH_USERNAME: aiden
-#      ME_CONFIG_BASICAUTH_PASSWORD: awesome
-#      ME_CONFIG_MONGODB_ADMINUSERNAME: root
-#      ME_CONFIG_MONGODB_ADMINPASSWORD: password
-#      ME_CONFIG_MONGODB_SERVER: mongodb
-
