+++ package.json
@@ -5,9 +5,9 @@
   "main": "src/server.ts",
   "scripts": {
     "build": "tsc",
-    "test": "tsc && sudo docker compose -f mongo-docker-compose.yml up -d && export NODE_ENV=test && mocha --recursive --exit",
-    "local": "tsc && sudo docker compose -f mongo-docker-compose.yml up -d && export MONGO_ENV=local && node dist/server.js",
-    "cloud": "tsc && export MONGO_ENV=cloud && node dist/server.js",
+    "test": "tsc && sudo docker compose -f postgres-docker-compose.yml up -d && export NODE_ENV=test && mocha --recursive --exit",
+    "local": "tsc && sudo docker compose -f postgres-docker-compose.yml up -d && export NODE_ENV=local && node dist/server.js",
+    "cloud": "tsc && export NODE_ENV=cloud && node dist/server.js",
     "lint": "eslint .",
     "lint_fix": "eslint --fix .",
     "generate_schema": "npx openapi-typescript ./hydRate.json -o ./schema.d.ts"
@@ -24,11 +24,12 @@
   "homepage": "https://github.com/awesomeaiden/hydrate-backend#readme",
   "dependencies": {
     "dotenv": "^16.3.1",
-    "express": "^4.18.1",
+    "express": "5.0.0-beta.3",
     "express-openapi-validator": "^5.1.6",
     "express-validator": "^6.14.3",
     "jsonwebtoken": "^9.0.2",
     "kysely": "^0.27.4",
+    "pg": "^8.12.0",
     "uuid": "^9.0.0",
     "validator": "^13.11.0"
   },
@@ -36,6 +37,7 @@
     "@types/chai": "^4.3.11",
     "@types/express": "^4.17.13",
     "@types/mocha": "^10.0.6",
+    "@types/pg": "^8.11.8",
     "@types/uuid": "^9.0.7",
     "@typescript-eslint/eslint-plugin": "^5.25.0",
     "@typescript-eslint/parser": "^5.25.0",
