+++ postgres-docker-compose.yml
@@ -1,0 +1,20 @@
+services:
+  db:
+    image: postgres
+    # restart: always
+    # set shared memory limit when using docker-compose
+    shm_size: 128mb
+    # or set shared memory limit when deploy via swarm stack
+    #volumes:
+    #  - type: tmpfs
+    #    target: /dev/shm
+    #    tmpfs:
+    #      size: 134217728 # 128*2^20 bytes = 128Mb
+    environment:
+      POSTGRES_PASSWORD: example
+
+  adminer:
+    image: adminer
+    # restart: always
+    ports:
+      - 8282:8080
