+++ src/db/database.ts
@@ -1,0 +1,21 @@
+import { Database } from "./types";
+import { Pool } from 'pg'
+import { Kysely, PostgresDialect } from 'kysely'
+
+const dialect = new PostgresDialect({
+  pool: new Pool({
+    database: 'postgres',
+    host: 'localhost',
+    user: 'postgres',
+    port: 5434,
+    max: 10
+  })
+});
+
+// Database interface is passed to Kysely's constructor, and from now on, Kysely
+// knows your database structure.
+// Dialect is passed to Kysely's constructor, and from now on, Kysely knows how
+// to communicate with your database.
+export const db = new Kysely<Database>({
+  dialect
+});
